For mercycorps/TolaActivity#2547, refactor and add bulk import file model.
