+++ indicators/views/__init__.py
@@ -44,3 +44,5 @@ from indicators.views.views_program import (
     api_program_page_indicator,
     results_framework_export,
 )
+
+from indicators.views.bulk_indicator_import_views import BulkImportIndicatorsView
